address deleted and merged into warehouse
