--- src/main/java/mn/mosc/project/domain/entity/address/Address.java
@@ -1,84 +1,0 @@
-package mn.mosc.project.domain.entity.address;
-
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBAttribute;
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBHashKey;
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBVersionAttribute;
-
-/**
- * created by ubulgan on 9/29/17
- */
-
-@DynamoDBTable(tableName = "ADDRESS")
-public class Address {
-    private String id;
-    private String country;
-    private String state;
-    private String city;
-    private String zipCode;
-    private String address;
-    private Long version;
-
-    @DynamoDBHashKey(attributeName = "id")
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    @DynamoDBAttribute(attributeName = "country")
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    @DynamoDBAttribute(attributeName = "state")
-    public String getState() {
-        return state;
-    }
-
-    public void setState(String state) {
-        this.state = state;
-    }
-
-    @DynamoDBAttribute(attributeName = "city")
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    @DynamoDBAttribute(attributeName = "zipCode")
-    public String getZipCode() {
-        return zipCode;
-    }
-
-    public void setZipCode(String zipCode) {
-        this.zipCode = zipCode;
-    }
-
-    @DynamoDBAttribute(attributeName = "address")
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    @DynamoDBVersionAttribute
-    public Long getVersion() {
-        return version;
-    }
-
-    public void setVersion(Long version) {
-        this.version = version;
-    }
-}
