+++ src/main/java/mn/mosc/project/domain/entity/inventory/Warehouse.java
@@ -1,7 +1,9 @@
 package mn.mosc.project.domain.entity.inventory;
 
-import com.amazonaws.services.dynamodbv2.datamodeling.*;
-import mn.mosc.project.domain.entity.address.Address;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBAttribute;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBHashKey;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBVersionAttribute;
 
 /**
  * created by loya on 9/29/17
@@ -9,7 +11,11 @@ import mn.mosc.project.domain.entity.address.Address;
 @DynamoDBTable(tableName = "WAREHOUSE")
 public class Warehouse {
     private String id;
-    private String addressId;
+    private String country;
+    private String state;
+    private String city;
+    private String zipCode;
+    private String address;
     private Long version;
 
     @DynamoDBHashKey(attributeName = "id")
@@ -21,14 +27,49 @@ public class Warehouse {
         this.id = id;
     }
 
-    @DynamoDBTyped(DynamoDBMapperFieldModel.DynamoDBAttributeType.M)
-    @DynamoDBAttribute(attributeName = "addressId")
-    public String getAddressId() {
-        return addressId;
+    @DynamoDBAttribute(attributeName = "country")
+    public String getCountry() {
+        return country;
     }
 
-    public void setAddressId(String address) {
-        this.addressId = addressId;
+    public void setCountry(String country) {
+        this.country = country;
+    }
+
+    @DynamoDBAttribute(attributeName = "state")
+    public String getState() {
+        return state;
+    }
+
+    public void setState(String state) {
+        this.state = state;
+    }
+
+    @DynamoDBAttribute(attributeName = "city")
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    @DynamoDBAttribute(attributeName = "zipCode")
+    public String getZipCode() {
+        return zipCode;
+    }
+
+    public void setZipCode(String zipCode) {
+        this.zipCode = zipCode;
+    }
+
+    @DynamoDBAttribute(attributeName = "address")
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
     }
 
     @DynamoDBVersionAttribute
